DjangoBlog/django_project/requirement.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.6
idna==2.8
Pillow==5.3.0
pytz==2018.5
requests==2.21.0
urllib3==1.24.1
django-crispy-forms==1.7.2